+++ IrcCommands/Default.py
@@ -1,5 +1,6 @@
 import datetime
 
+#Called when pm'd with anything that's not a function
 def run(user, msg, ircClient):
 	now = datetime.datetime.now()
 	time = now.strftime('%r')
