+++ IrcCommands/Discord.py
@@ -1,3 +1,5 @@
+#Called when pm'd with
+#	!discord
 def run(user, msg, ircClient):
     ircClient.msg(user, 'The place to talk about the bot or just chat: https://discord.gg/hKXQdm2')
     print(f'Printed discord invite for {user}.')
