+++ IrcCommands/With.py
@@ -1,8 +1,10 @@
 import re
-import config
-import pp
-from roundString import roundString as roundString
+from Utils import config
+from Utils import pp
+from Utils import roundString as roundString
 
+#Called when pm'd with
+#	!with <acc> <misses>
 def run(user, msg, ircClient):
 	irc = ircClient
 	conf = config.Config('bot.conf')
